mm-pp-daemon: Enable mm-pp-daemon.
The mm-pp-daemon is responsible for setting color settings
and cabl as well. There is an pp_calib_data file needed to
get the color temperature feature working. The one within
this commit is taken from
pp_calib_data_sharp_r63315_1080p_video_mode_dsi_panel.xml
and has to be adapted for the current display with appropriate
values. As start it seems to give good results.

The shims are extended to load libdisplayconfig.so when loading
the libqdutils.so. This is needed due a recent change, where the
refreshScreen() function was moved to a dedicated library. To solve
the issue, load the libdisplayconfig to let these functions be visible
to dlsym().

Enable CABL by setting appropriate arguments to the mm-pp-daemon.
This has to be reverted in future and replaced by a proper implementation
of a derrived device specific LiveDisplayMM controller.
Also the system property is set to ro.qualcomm.cabl=2 which maches the stock
values.

Change-Id: I81c3884fa15a1f65f74427feb4ccbadc6e394faa